Split the component for better readability
