+++ src/components/tasks/card/index.tsx
@@ -31,9 +31,7 @@ import { useEditMode } from '@/hooks/useEditMode';
 import { useGetUsersByUsernameQuery } from '@/app/services/usersApi';
 import { ConditionalLinkWrapper } from './ConditionalLinkWrapper';
 import { isNewCardDesignEnabled } from '@/constants/FeatureFlags';
-import SuggestionBox from '../SuggestionBox/SuggestionBox';
-import { userDataType } from '@/interfaces/user.type';
-import { GithubInfo } from '@/interfaces/suggestionBox.type';
+import Suggestions from './Suggestions';
 
 type Props = {
     content: task;
@@ -41,8 +39,6 @@ type Props = {
     onContentChange?: (changeId: string, changeObject: object) => void;
     updateTask?: (taskId: string, details: updateTaskDetails) => void;
 };
-
-let timer: NodeJS.Timeout;
 
 const Card: FC<Props> = ({
     content,
@@ -77,10 +73,6 @@ const Card: FC<Props> = ({
         itemId: cardDetails.id,
     });
     const [deleteTaskTagLevel] = useDeleteTaskTagLevelMutation();
-
-    const [isLoadingSuggestions, setIsLoadingSuggestions] =
-        useState<boolean>(false);
-    const [suggestions, setSuggestions] = useState<GithubInfo[]>([]);
     const [assigneeName, setAssigneeName] = useState<string>('');
     const inputRef = useRef<HTMLInputElement>(null);
 
@@ -394,43 +386,6 @@ const Card: FC<Props> = ({
         setShowSuggestion(false);
     };
 
-    const fetchUsers = async (e: string) => {
-        if (!e) return;
-        setIsLoadingSuggestions(true);
-
-        const url = `${process.env.NEXT_PUBLIC_BASE_URL}/users?search=${e}`;
-        try {
-            const { requestPromise } = fetch({ url });
-            const users = await requestPromise;
-            const usersData = users.data.users;
-            const suggestedUsers: GithubInfo[] = [];
-            usersData.map((data: userDataType) => {
-                suggestedUsers.push({
-                    github_id: data.username,
-                    profileImageUrl: data?.picture?.url
-                        ? data.picture.url
-                        : placeholderImageURL,
-                });
-            });
-
-            setSuggestions(suggestedUsers);
-            setIsLoadingSuggestions(false);
-        } catch (error: any) {
-            setIsLoadingSuggestions(false);
-            toast(ERROR, error.message);
-        }
-    };
-
-    const debounce = (fn: (e: string) => void, delay: number) => {
-        return function (e: string) {
-            clearTimeout(timer);
-            setSuggestions([]);
-            timer = setTimeout(() => {
-                fn(e);
-            }, delay);
-        };
-    };
-
     // show redesign only on dev
     if (isNewCardDesignEnabled)
         return (
@@ -506,36 +461,14 @@ const Card: FC<Props> = ({
                         </span>
                         {shouldEdit ? (
                             isUserAuthorized && (
-                                <div className={classNames.suggestionDiv}>
-                                    <input
-                                        ref={inputRef}
-                                        value={assigneeName}
-                                        className={classNames.cardStrongFont}
-                                        onKeyDown={(e) => {
-                                            handleChange(e, 'assignee');
-                                        }}
-                                        onChange={(e) => {
-                                            handleAssignment(e);
-                                            debounce(
-                                                fetchUsers,
-                                                400
-                                            )(e.target.value);
-                                        }}
-                                        role="button"
-                                        tabIndex={0}
-                                    />
-
-                                    {isLoadingSuggestions ? (
-                                        <Loader />
-                                    ) : (
-                                        showSuggestion && (
-                                            <SuggestionBox
-                                                suggestions={suggestions}
-                                                onSelectAssignee={handleClick}
-                                            />
-                                        )
-                                    )}
-                                </div>
+                                <Suggestions
+                                    assigneeName={assigneeName}
+                                    showSuggestion={showSuggestion}
+                                    handleAssignment={handleAssignment}
+                                    handleClick={handleClick}
+                                    handleChange={handleChange}
+                                    ref={inputRef}
+                                />
                             )
                         ) : (
                             <span className={classNames.cardStrongFont}>
@@ -667,40 +600,14 @@ const Card: FC<Props> = ({
                             </span>
                             {shouldEdit ? (
                                 isUserAuthorized && (
-                                    <div className={classNames.suggestionDiv}>
-                                        <input
-                                            ref={inputRef}
-                                            value={assigneeName}
-                                            className={
-                                                classNames.cardStrongFont
-                                            }
-                                            onKeyDown={(e) => {
-                                                handleChange(e, 'assignee');
-                                            }}
-                                            onChange={(e) => {
-                                                handleAssignment(e);
-                                                debounce(
-                                                    fetchUsers,
-                                                    400
-                                                )(e.target.value);
-                                            }}
-                                            role="button"
-                                            tabIndex={0}
-                                        />
-
-                                        {isLoadingSuggestions ? (
-                                            <Loader />
-                                        ) : (
-                                            showSuggestion && (
-                                                <SuggestionBox
-                                                    suggestions={suggestions}
-                                                    onSelectAssignee={
-                                                        handleClick
-                                                    }
-                                                />
-                                            )
-                                        )}
-                                    </div>
+                                    <Suggestions
+                                        assigneeName={assigneeName}
+                                        showSuggestion={showSuggestion}
+                                        handleAssignment={handleAssignment}
+                                        handleClick={handleClick}
+                                        handleChange={handleChange}
+                                        ref={inputRef}
+                                    />
                                 )
                             ) : (
                                 <span className={classNames.cardStrongFont}>
